requirements.txt
pip~=24.0
wheel
Cython
transformers==4.40.2
youtokentome
huggingface-hub==0.23.2
nemo-toolkit[all]
pytorch-lightning==2.0.1
pandas
hydra-core
sentencepiece
git+https://github.com/gburlet/YouTokenToMe.git@dependencies
braceexpand
webdataset
h5py
psutil
ijson
sacremoses

matplotlib
sacrebleu
rouge_score
einops
jieba
opencc
pangu

torchtext